backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.3.0
psycopg2-binary==2.9.9
requests==2.31.0
gunicorn==21.2.0